pyproject.toml
[project]
name = "congestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Thiago", email = "bomfimdev@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.5.2",
    "linearmodels>=7.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.9",
    "numpy>=2.2.6",
    "plotly>=6.7.0",
    "polars>=1.40.1",
    "pyarrow>=24.0.0",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "scipy>=1.15.3",
    "statsmodels>=0.14.6",
    "streamlit>=1.57.0",
    "tenacity>=9.1.4",
]

[project.scripts]
congestion = "congestion:main"

[build-system]
requires = ["uv_build>=0.11.8,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]
